Rename IterTools to IteratorTools to avoid problems with mac
Conflicts:
	src/python/WMComponent/DBS3Buffer/MySQL/DBSBufferFiles/AddRunLumi.py
	src/python/WMCore/BossAir/Plugins/PyCondorPlugin.py
	src/python/WMCore/BossAir/Plugins/SimpleCondorPlugin.py
	src/python/WMCore/WMBS/MySQL/Files/AddRunLumi.py
