+++ src/python/WMComponent/DBS3Buffer/MySQL/DBSBufferFiles/AddRunLumi.py
@@ -6,8 +6,6 @@ MySQL implementation of AddRunLumi
 """
 
 
-
-
 from WMCore.Database.DBFormatter import DBFormatter
 
 class AddRunLumi(DBFormatter):
